promote speed field in crac json to remedial action level
