+++ data/crac/crac-api/src/main/java/com/farao_community/farao/data/crac_api/RemedialActionAdder.java
@@ -16,6 +16,8 @@ public interface RemedialActionAdder<T extends RemedialActionAdder<T>> extends I
 
     T withOperator(String operator);
 
+    T withSpeed(Integer speed);
+
     OnInstantAdder<T> newOnInstantUsageRule();
 
     OnContingencyStateAdder<T> newOnContingencyStateUsageRule();
