+++ data/crac/crac-api/src/main/java/com/farao_community/farao/data/crac_api/range_action/PstRangeActionAdder.java
@@ -24,8 +24,6 @@ public interface PstRangeActionAdder extends RemedialActionAdder<PstRangeActionA
 
     PstRangeActionAdder withGroupId(String groupId);
 
-    PstRangeActionAdder withSpeed(Integer speed);
-
     PstRangeActionAdder withInitialTap(int initialTap);
 
     PstRangeActionAdder withTapToAngleConversionMap(Map<Integer, Double> tapToAngleConversionMap);
