+++ data/crac/crac-api/src/main/java/com/farao_community/farao/data/crac_api/range_action/RangeAction.java
@@ -54,8 +54,4 @@ public interface RangeAction<T extends RangeAction<T>> extends RemedialAction<T>
      */
     Optional<String> getGroupId();
 
-    /**
-     * Get the speed of the Range Action, i.e the time it takes to trigger.
-     */
-    Optional<Integer> getSpeed();
 }
