+++ data/crac/crac-api/src/main/java/com/farao_community/farao/data/crac_api/range_action/StandardRangeActionAdder.java
@@ -20,6 +20,4 @@ public interface StandardRangeActionAdder<T extends StandardRangeActionAdder<T>>
 
     T withInitialSetpoint(double initialSetpoint);
 
-    T withSpeed(Integer speed);
-
 }
